+++ CloudStorage/client/pom.xml
@@ -40,6 +40,13 @@
             <version>16</version>
         </dependency>
 
+        <dependency>
+            <groupId>com.pae</groupId>
+            <artifactId>common</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+
     </dependencies>
 
     <build>
