--- CloudStorage/client/src/main/java/com/pae/cloudstorage/client/common/CallBack.java
@@ -1,5 +1,0 @@
-package com.pae.cloudstorage.client.common;
-
-public interface CallBack {
-    void call(Object... args);
-}
