+++ CloudStorage/common/src/main/java/com/pae/cloudstorage/common/CallBack.java
@@ -1,0 +1,5 @@
+package com.pae.cloudstorage.common;
+
+public interface CallBack {
+    void call(Object... args);
+}
