+++ CloudStorage/common/src/main/java/module-info.java
@@ -1,0 +1,3 @@
+module common {
+    exports com.pae.cloudstorage.common;
+}
