--- CloudStorage/server/src/main/java/com/pae/cloudstorage/server/common/CallBack.java
@@ -1,6 +1,0 @@
-package com.pae.cloudstorage.server.common;
-
-@FunctionalInterface
-public interface CallBack {
-    void call(Object... args);
-}
